Gave menu its own Class and put the eventhandler in there;
